# public/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Star Wars API + Laravel Livewire Code10 Coding Exercise",
        "version": "1.0",
        "description": "Api specification and documentation for how this app interacts with StarWars API.",
        "termsOfService": "",
        "license": null,
        "contact": {
            "name": "JD Hawkins",
            "email": "itssimplescience@gmail.com"
        },
        "x-logo": {
            "url": "https:\/\/code-10-assignment.itssimplescience.com:8890\/assets\/media\/svg\/humans\/custom-1.svg",
            "altText": "Star Wars API + Laravel Livewire Code10 Coding Exercise",
            "backgroundColor": ""
        }
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
...
    },
    "servers": [
        {
            "url": "https:\/\/code-10-assignment.itssimplescience.com:8890",
            "description": "Documentation generator server."
        }
    ],
    "paths": {
        "\/api\/v1\/people": {
            "get": {
                "tags": [
                    "Person API Specification"
                ],
                "operationId": "List persons",
                "description": "Fetch paginated collection of persons\nconvert items into a PersonResourceCollection",
                "parameters": [
                    {
                        "in": "query",
                        "name": "search",
                        "description": "allows you to search results by provided param.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Darth Maul"
                        }
                    },
                    {
                        "in": "query",
                        "name": "page",
                        "description": "allows you to set the current page.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "type": "object",
                                    "example": {
                                        "code": 200,
                                        "status": "OK",
                                        "message": "OK",
                                        "data": [
                                            {
                                                "format": null,
                                                "url": "https:\/\/swapi.dev\/api\/people\/44\/",
                                                "image": "http:\/\/localhost\/assets\/media\/avatars\/300-8.jpg",
                                                "homeWorldImage": "http:\/\/localhost\/assets\/media\/stock\/600x400\/img-36.jpg",
                                                "name": "Darth Maul",
                                                "height": 175,
                                                "mass": 80,
                                                "hairColor": "none",
                                                "skinColor": "red",
                                                "eyeColor": "yellow",
                                                "birthYear": "54BBY",
                                                "gender": "male",
                                                "homeWorld": "https:\/\/swapi.dev\/api\/planets\/36\/",
                                                "createdOn": "2014-12-19T18:00:41.929000Z",
                                                "editedOn": "2014-12-20T21:17:50.403000Z",
                                                "noOfFilms": 1,
                                                "noOfStarShips": 1,
                                                "noOfVehicles": 1
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    }
                },
                "x-code-samples": [
                    {
                        "lang": "Bash",
                        "source": "curl -X GET -G \"https:\/\/code-10-assignment.itssimplescience.com:8890\/api\/v1\/people\" \n"
                    },
                    {
                        "lang": "Javascript",
                        "source": "const url = new URL(\"https:\/\/code-10-assignment.itssimplescience.com:8890\/api\/v1\/people\");\n\nlet params = {\n    \"search\": \"Darth Maul\",\n    \"page\": \"1\",\n};\n\nObject.keys(params).forEach(key => url.searchParams.append(key, params[key]));\n\nlet headers = {\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nfetch(url, {\n    method: \"GET\",\n    headers: headers,\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"
                    },
                    {
                        "lang": "PHP",
                        "source": "$headers = [\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n$query = [\n    \"search\" => \"Darth Maul\",\n    \"page\" => \"1\",\n];\n\n\n$ch = curl_init(\"https:\/\/code-10-assignment.itssimplescience.com:8890\/api\/v1\/people?\" . http_build_query($data));\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"
                    }
                ]
            }
        },
        "\/api\/v1\/people\/{personId}": {
            "get": {
                "tags": [
                    "Person API Specification"
                ],
                "operationId": "Show person",
                "description": "Fetch the details of a single PersonResource\nfetch all StarShipResource related to person",
                "parameters": [
                    {
                        "in": "path",
                        "name": "personId",
                        "description": "The id of the person object.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 17
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "type": "object",
                                    "example": {
                                        "code": 200,
                                        "status": "OK",
                                        "message": "OK",
                                        "data": {
                                            "format": null,
                                            "url": "https:\/\/swapi.dev\/api\/people\/1\/",
                                            "image": "http:\/\/localhost\/assets\/media\/avatars\/300-8.jpg",
                                            "homeWorldImage": "http:\/\/localhost\/assets\/media\/stock\/600x400\/img-48.jpg",
                                            "name": "Luke Skywalker",
                                            "height": 172,
                                            "mass": 77,
                                            "hairColor": "blond",
                                            "skinColor": "fair",
                                            "eyeColor": "blue",
                                            "birthYear": "19BBY",
                                            "gender": "male",
                                            "homeWorld": "https:\/\/swapi.dev\/api\/planets\/1\/",
                                            "createdOn": "2014-12-09T13:50:51.644000Z",
                                            "editedOn": "2014-12-20T21:17:56.891000Z",
                                            "noOfFilms": 4,
                                            "noOfStarShips": 2,
                                            "noOfVehicles": 2,
                                            "starships": [
                                                {
                                                    "image": "http:\/\/localhost\/assets\/media\/stock\/900x600\/4.jpg",
                                                    "name": "X-wing",
                                                    "model": "T-65 X-wing",
                                                    "manufacturer": "Incom Corporation",
                                                    "cost_in_credits": 149999,
                                                    "length": 12,
                                                    "max_atmosphering_speed": 1050,
                                                    "crew": 1,
                                                    "passengers": 0,
                                                    "cargo_capacity": 110,
                                                    "consumables": "1 week",
                                                    "hyperdrive_rating": 1,
                                                    "starship_class": "Starfighter",
                                                    "created": "2014-12-12T11:19:05.340000Z",
                                                    "edited": "2014-12-20T21:23:49.886000Z",
                                                    "url": "https:\/\/swapi.dev\/api\/starships\/12\/"
                                                },
                                                {
                                                    "image": "http:\/\/localhost\/assets\/media\/stock\/900x600\/67.jpg",
                                                    "name": "Imperial shuttle",
                                                    "model": "Lambda-class T-4a shuttle",
                                                    "manufacturer": "Sienar Fleet Systems",
                                                    "cost_in_credits": 240000,
                                                    "length": 20,
                                                    "max_atmosphering_speed": 850,
                                                    "crew": 6,
                                                    "passengers": 20,
                                                    "cargo_capacity": 80000,
                                                    "consumables": "2 months",
                                                    "hyperdrive_rating": 1,
                                                    "starship_class": "Armed government transport",
                                                    "created": "2014-12-15T13:04:47.235000Z",
                                                    "edited": "2014-12-20T21:23:49.900000Z",
                                                    "url": "https:\/\/swapi.dev\/api\/starships\/22\/"
                                                }
                                            ]
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "x-code-samples": [
                    {
                        "lang": "Bash",
                        "source": "curl -X GET -G \"https:\/\/code-10-assignment.itssimplescience.com:8890\/api\/v1\/people\/{personId}\" \n"
                    },
                    {
                        "lang": "Javascript",
                        "source": "const url = new URL(\"https:\/\/code-10-assignment.itssimplescience.com:8890\/api\/v1\/people\/{personId}\");\n\nlet headers = {\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nfetch(url, {\n    method: \"GET\",\n    headers: headers,\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"
                    },
                    {
                        "lang": "PHP",
                        "source": "$headers = [\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n\n$ch = curl_init(\"https:\/\/code-10-assignment.itssimplescience.com:8890\/api\/v1\/people\/{personId}\");\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"
                    }
                ]
            }
        },
...
    },
...
}